+++ CHANGELOG.md
@@ -6,6 +6,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.17] - 2020-05-13
+### Fixed
+* [`useForm`, `useValidation`] Fixed an issues with the `errors` types.
+
+## [0.0.16] - 2020-04-04
+### Added
+* [`useForm`, `useTouch`] Adds a isDirty field that is set to true when any field is touched.
+
+## [0.0.15] - 2020-03-22
+### Fixed
+* [`useForm`] Fixed an issue where `onSubmit` was being called with the current values.
+
+## [0.0.14] - 2020-03-22
+### Added
+* [`useSubmission`] `onSubmit` will be called with the current Event if `submit` is called with an event.
+* [`useForm`] `onSubmit` the second argument will be the event if `submit` is called with an event.
+
+### Removed
+* [`useValidation`] Removed `isValidating` from the hook.
+A hook to track if a function is currently being called will be added in a later update.
+
 ## [0.0.13] - 2020-03-08
 ### Added
 * Typescript definition files